chg: basic pixel object picking

+++ src/context.h
@@ -18,7 +18,7 @@ struct Context
 
   // Editor state
   uint64_t selAssetUUID{0};
-  uint64_t selObjectUUID{0};
+  uint32_t selObjectUUID{0};
   bool isBuildOrRunning{false};
 };
